fix(IrbParams): fix lowestlow type

+++ model/strategy/params/strategy_params.py
@@ -12,7 +12,7 @@ class TrendParams(BaseModel):
     trend: bool = False
 
 class IrbParams(BaseModel):
-    lowestlow: float = 1
+    lowestlow: int = 1
     payoff: float = 2
     tick_size: float = 0.1
     wick_percentage: float = 0.45
